refactor: Move configurations to utils module
