+++ airflow_dbt_python/utils/enums.py
@@ -1,14 +1,23 @@
-"""Enumerations with allowed set of values."""
+"""Enumeration subclasses used in dbt configurations."""
 from enum import Enum
+from typing import Type, TypeVar
+
+FromStrEnumSelf = TypeVar("FromStrEnumSelf", bound="FromStrEnum")
 
 
 class FromStrEnum(Enum):
     """Access enum variants with strings ensuring uppercase."""
 
     @classmethod
-    def from_str(cls, s: str):
+    def from_str(cls: Type[FromStrEnumSelf], name: str) -> FromStrEnumSelf:
         """Instantiate an Enum from a string."""
-        return cls[s.replace("-", "_").upper()]
+        return cls[name.replace("-", "_").upper()]
+
+    def __eq__(self, other):
+        """Override equality for string comparison."""
+        if isinstance(other, str):
+            return other.upper() == self.name
+        return Enum.__eq__(self, other)
 
 
 class LogFormat(FromStrEnum):
@@ -26,9 +35,3 @@ class Output(FromStrEnum):
     NAME = "name"
     PATH = "path"
     SELECTOR = "selector"
-
-    def __eq__(self, other):
-        """Override equality for string comparison."""
-        if isinstance(other, str):
-            return other.upper() == self.name
-        return Enum.__eq__(self, other)
